requirements.txt
livekit>=0.17.0
livekit-agents>=1.2.0
livekit-plugins-openai>=0.8.0
livekit-plugins-silero>=0.6.0
python-dotenv>=1.0.0